Module google_videointelligence1::api

Structs

CloudVideoIntelligence
Central instance to access all CloudVideoIntelligence related resource activities

GoogleCloudVideointelligenceV1_AnnotateVideoRequest
Video annotation request.

GoogleCloudVideointelligenceV1_ExplicitContentDetectionConfig
Config for EXPLICIT_CONTENT_DETECTION.

GoogleCloudVideointelligenceV1_FaceDetectionConfig
Config for FACE_DETECTION.

GoogleCloudVideointelligenceV1_LabelDetectionConfig
Config for LABEL_DETECTION.

GoogleCloudVideointelligenceV1_ObjectTrackingConfig
Config for OBJECT_TRACKING.

GoogleCloudVideointelligenceV1_PersonDetectionConfig
Config for PERSON_DETECTION.

GoogleCloudVideointelligenceV1_ShotChangeDetectionConfig
Config for SHOT_CHANGE_DETECTION.

GoogleCloudVideointelligenceV1_SpeechContext
Provides "hints" to the speech recognizer to favor specific words and phrases in the results.

GoogleCloudVideointelligenceV1_SpeechTranscriptionConfig
Config for SPEECH_TRANSCRIPTION.

GoogleCloudVideointelligenceV1_TextDetectionConfig
Config for TEXT_DETECTION.

GoogleCloudVideointelligenceV1_VideoContext
Video context and/or feature-specific parameters.

GoogleCloudVideointelligenceV1_VideoSegment
Video segment.

GoogleLongrunning_CancelOperationRequest
The request message for Operations.CancelOperation.

GoogleLongrunning_ListOperationsResponse
The response message for Operations.ListOperations.

GoogleLongrunning_Operation
This resource represents a long-running operation that is the result of a network API call.

GoogleProtobuf_Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

GoogleRpc_Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

OperationMethods
A builder providing access to all methods supported on operation resources. It is not used directly, but through the CloudVideoIntelligence hub.

OperationProjectLocationOperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

OperationProjectLocationOperationDeleteCall
Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

OperationProjectLocationOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectLocationOperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

ProjectLocationOperationDeleteCall
Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

ProjectLocationOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectLocationOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the CloudVideoIntelligence hub.

VideoAnnotateCall
Performs asynchronous video annotation. Progress and results can be retrieved through the google.longrunning.Operations interface. Operation.metadata contains AnnotateVideoProgress (progress). Operation.response contains AnnotateVideoResponse (results).

VideoMethods
A builder providing access to all methods supported on video resources. It is not used directly, but through the CloudVideoIntelligence hub.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.